+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Infrastructure
 * Disable changelog verification for dependabot PRs
 
+### Fixed
+* Reconcile missing Application Key caused by the resource drift
+* Fix reconciliation of missing Bucket caused by the resource drift
+
 ## [0.8.4] - 2023-03-13
 
 ### Infrastructure